+++ svc_model.py
@@ -7,8 +7,22 @@ import cv2
 from sklearn.preprocessing import LabelEncoder
 from sklearn.svm import SVC
 
-from extract import calc_embs
+def predict(filename,calc_embs,encoder,clf):
+    embs,boxs,imgs = calc_embs([filename])
 
+    if embs is not None:
+        copy = imgs[0].copy()
+        embs = np.concatenate(embs)
+
+        labels = encoder.inverse_transform(clf.predict([embs]))
+
+        print(labels)
+        # x,y,w,h = boxs[0]
+
+        return labels[0],boxs[0],copy
+    else:
+        copy = cv2.imread(filename,0)
+        return None,None,copy
 
 def train():
     parser = argparse.ArgumentParser(description='Classification with SVM.')
@@ -24,8 +38,14 @@ def train():
     args = parser.parse_args()
 
     embdding_path = args.embdding
+
     max_num_img = args.num
+    print("max_num_img : ",max_num_img)
+
     folder_model = args.model
+    if not os.path.exists(folder_model):
+        os.mkdir(folder_model)
+
     mode = args.mode
     # filename = args.image_path
 
@@ -42,6 +62,7 @@ def train():
         
         embs = np.concatenate(embs)
         print(embs.shape)
+        print(labels)
         le = LabelEncoder().fit(labels)
         y = le.transform(labels)
         print(y)
@@ -55,6 +76,9 @@ def train():
             pickle.dump(le,f)
 
     elif mode == "test":
+        from extract import calc_embs
+        wd = cv2.namedWindow("",cv2.WINDOW_FREERATIO)
+
         with open(folder_model+"/svc_model.pickle","rb") as f:
             clf = pickle.load(f)
 
@@ -62,39 +86,53 @@ def train():
             le = pickle.load(f)
 
         while True:
-            print("Enter your image file...")
-            filename = input()
+            filename = ""
+            timeout = 100
+
+            filename = input("Enter your image file : ")
             if filename == "":
-                print("Do you want to quit(y/n)?")
-                msg = input()
+                msg = input("Do you want to quit(y/n)? : ")
                 if msg == "y":
                     print("Byebye , see you again.")
                     break
                 elif msg == "n":
-                    print("Enter your image file...")
-                    filename = input()
-                    pass
+                    filename = input("Enter your image file : ")
 
-            embs,boxs,imgs = calc_embs([filename])
+            if os.path.isfile(filename) and os.path.exists(filename):
 
-            embs = np.concatenate(embs)
+                label , box , copy = predict(filename,calc_embs,le,clf)
+                if label is not None:
+                    x,y,w,h = box
+                    cv2.rectangle(copy, (x,y), (x++w,y+h), (255,0,0),2)
+                    cv2.putText(copy,label,(x,y),cv2.FONT_HERSHEY_COMPLEX,1,(255,0,0),2)
+                cv2.imshow(wd,copy)
+                cv2.waitKey(0)
 
-            labels = le.inverse_transform(clf.predict([embs]))
+                cv2.imwrite("output.jpg",copy)
+                # labels = ",".join(labels)
 
-            print(labels)
+                with open("labels.txt","w") as of:
+                    of.write(label)
 
-            copy = imgs[0].copy()
-            x,y,w,h = boxs[0]
-            cv2.rectangle(copy, (x,y), (x++w,y+h), (255,0,0),2)
-            cv2.putText(copy,labels[0],(x,y),cv2.FONT_HERSHEY_COMPLEX,1,(255,0,0),2)
-            cv2.imshow("",copy)
-            cv2.waitKey(0)
+            elif os.path.isdir(filename) and os.path.exists(filename):
+                timeout = input("Enter timeout : ")
+                try:
+                    timeout = int(timeout)
+                except:
+                    timeout = 100
 
-            cv2.imwrite("output.jpg",copy)
-            labels = ",".join(labels)
+                for f in os.listdir(filename):
+                    file = os.path.join(filename,f)
+                    label , box , copy = predict(file,calc_embs,le,clf)
 
-            with open("labels.txt","w") as of:
-                of.write(labels)
+                    if label is not None:
+                        x,y,w,h = box
+                        cv2.rectangle(copy, (x,y), (x++w,y+h), (255,0,0),2)
+                        cv2.putText(copy,label,(x,y),cv2.FONT_HERSHEY_COMPLEX,1,(255,0,0),2)
+                    cv2.imshow(wd,copy)
+                    cv2.waitKey(timeout)
+
+        cv2.destroyAllWindows()
 
 if __name__ == "__main__":
     train()
